src/restarter/pip-requirements.txt
psutil==5.9.8
dbus-python==1.3.2
PySimpleGUI==4.60.5
